debug make complex

+++ utils/numeric.py
@@ -14,7 +14,7 @@ def make_complex(x: Union[np.ndarray, torch.Tensor]) -> torch.Tensor:
     compl = make_tensor(x) 
     if not torch.is_complex(compl): 
         compl = torch.view_as_complex(compl) if compl.shape[-1] == 2 else torch.complex(compl, torch.zeros_like(compl))
-    return compl
+    return compl.to(torch.cfloat)
 
 
 def make_real(x: Union[np.ndarray, torch.Tensor]) -> torch.Tensor:
